third_party/blink/web_tests/fast/css/font-face-cache-bug-expected.txt

Test for WebKit bug 73520

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.getComputedStyle(testDiv, 0).width is "48px"
PASS successfullyParsed is true

TEST COMPLETE

aaa
